+++ docs/changelog.md
@@ -9,6 +9,7 @@
 + Add "navigation toolbar position" option.
 + Add "auto remove empty links after deleted points" option. (github issue #23)
 + Move background option into project file. (github issue #26)
++ The background image can be set by dragging in.
 
 **Development**: